% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createData.R
\name{createDataExt}
\alias{createDataExt}
\title{Simulate test data}
\usage{
createDataExt(replicates = 1, sampleSize = 10, extPredictors = NULL,
  intercept = 0, fixedEffects = 1, quadraticFixedEffects = NULL,
  numGroups = 10, randomEffectVariance = 1, overdispersion = 0,
  family = gaussian(), scale = 1, cor = 0, roundPoissonVariance = NULL,
  pZeroInflation = 0, binomialTrials = 1, temporalAutocorrelation = 0,
  spatialAutocorrelation = 0, factorResponse = F)
}
\arguments{
\item{replicates}{number of datasets to create}

\item{sampleSize}{sample size of the dataset}

\item{extPredictors}{dataframe of environmental data generated externally}

\item{intercept}{intercept (linear scale)}

\item{fixedEffects}{vector of fixed effects (linear scale)}

\item{quadraticFixedEffects}{vector of quadratic fixed effects (linear scale)}

\item{numGroups}{number of groups for the random effect}

\item{randomEffectVariance}{variance of the random effect (intercept)}

\item{overdispersion}{if this is a numeric value, it will be used as the sd of a random normal variate that is added to the linear predictor. Alternatively, a random function can be provided that takes as input the linear predictor.}

\item{family}{family}

\item{scale}{scale if the distribution has a scale (e.g. sd for the Gaussian)}

\item{cor}{correlation between predictors}

\item{roundPoissonVariance}{if set, this creates a uniform noise on the possion response. The aim of this is to create heteroscedasticity}

\item{pZeroInflation}{probability to set any data point to zero}

\item{binomialTrials}{Number of trials for the binomial. Only active if family == binomial}

\item{temporalAutocorrelation}{strength of temporalAutocorrelation}

\item{spatialAutocorrelation}{strength of spatial Autocorrelation}

\item{factorResponse}{should the response be transformed to a factor (intended to be used for 0/1 data)}
}
\description{
This function creates synthetic dataset with various problems such as overdispersion, zero-inflation, etc. and was modified to use also externally generated environmental data.
}
\note{
The basic structure of this function was taken from \code{\link[DHARMa]{createData}} in the package DHARMa by Florian Hartig.
}
\examples{
require(lme4)


# complex environment

testData = createDataExt(sampleSize = 2000, intercept = 5, fixedEffects = c(2, 2, 0), cor = 0.5)

pairs(testData[, grepl("Environment", names(testData))])

fittedModel <- lmer(observedResponse ~ Environment1 + Environment2 + Environment3 + (1|group), data = testData)
summary(fittedModel)



# External predictors

nvar <- 4
mu <- sample(seq(0.01, 0.5, length.out = 50), nvar, replace = TRUE)
data_sim <- corrEnv(n = 2000, nvar = nvar, ngrad = 3, mu = mu, rho = 0.9, rho.non.corr = 0)

testData = createDataExt(sampleSize = 2000, extPredictors = data_sim[[2]], fixedEffects = c(2, 2, 1, 0))

hist(testData$observedResponse)
pairs(testData[, grepl("Environment", names(testData))])

fittedModel <- lmer(observedResponse ~ Environment1 + Environment2 + Environment3 + Environment4 + (1|group), data = testData)
summary(fittedModel)



### FROM DHARMa



# with zero-inflation

testData = createDataExt(sampleSize = 500, intercept = 2, fixedEffects = c(1),
                      overdispersion = 0, family = poisson(), quadraticFixedEffects = c(-3),
                      randomEffectVariance = 0, pZeroInflation = 0.6)

par(mfrow = c(1,2))
plot(testData$Environment1, testData$observedResponse)
hist(testData$observedResponse)


# binomial with multiple trials

testData = createDataExt(sampleSize = 40, intercept = 2, fixedEffects = c(1),
                      overdispersion = 0, family = binomial(), quadraticFixedEffects = c(-3),
                      randomEffectVariance = 0, binomialTrials = 20)

plot(observedResponse1 / observedResponse0 ~ Environment1, data = testData, ylab = "Proportion 1")


# spatial / temporal correlation

testData = createDataExt(sampleSize = 100, family = poisson(), spatialAutocorrelation = 3,
                      temporalAutocorrelation = 3)

par(mfrow = c(1,2))
plot(log(observedResponse) ~ time, data = testData)
plot(log(observedResponse) ~ x, data = testData)










}
\author{
Lisa Huelsmann
}
