% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{shannon_diversity}
\alias{shannon_diversity}
\title{Simple Shannon Diversity Calculator}
\usage{
shannon_diversity(counts)
}
\arguments{
\item{counts}{Numeric vector of species counts}
}
\value{
Shannon diversity index
}
\description{
Simple Shannon Diversity Calculator
}
\examples{
shannon_diversity(c(10, 15, 8, 12, 5))
}
