% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoteach-package.R
\docType{package}
\name{ecoteach-package}
\alias{ecoteach}
\alias{ecoteach-package}
\title{ecoteach: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
