% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple_data}
\alias{simple_data}
\title{Simple Ecological Data}
\format{
A data frame with 20 rows and 3 columns
}
\usage{
simple_data
}
\description{
Simple Ecological Data
}
\examples{
data(simple_data)
head(simple_data)
}
\keyword{datasets}
